Added notification styles
Added notification style options for `success` and `error`. 
The default is `info` which is black.

Setting a style option of `success` will make the notification background green (50% opacity).
Setting a style option of `error` will make the notification background red (50% opacity).

This is useful for when you want to show successful and erroneous login messages for example.